Few fixes

+++ App/Timer/Timer.cpp
@@ -13,7 +13,7 @@ void Timer::Mark() noexcept
 	std::chrono::steady_clock::time_point currentTime = std::chrono::steady_clock::now();
 
 	pFrameTime = currentTime - pLast;
-	pFPS = 1 / pFrameTime.count();
+	pFPS = static_cast<int>(1 / pFrameTime.count());
 
 	pLast = currentTime;
 }
@@ -21,14 +21,14 @@ void Timer::Mark() noexcept
 bool Timer::ShowFrame() noexcept
 {
 	std::chrono::steady_clock::time_point currentTime = std::chrono::steady_clock::now();
-	if (1 / (float)pFPS_max <= std::chrono::duration<double>(currentTime - pLast).count())
+	if (1 / (float)pFPS_max <= std::chrono::duration<float>(currentTime - pLast).count())
 	{
 		return true;
 	}
 	return false;
 }
 
-double Timer::GetTimePerFrame() noexcept
+float Timer::GetTimePerFrame() noexcept
 {
 	return pFrameTime.count();
 }
